x-button now throws an event
This is a workaround to consider if the 'x' button is pressed to close the input without an submit. The submit originally calls the callback and executes the InputsService.CloseInput() method after that. The 'x' button calls the InputsService.CloseInput() directly. So we need to check if the callback was already called by the submit otherwise the trigger was the 'x' button then we need to call the callback now.